+++ CHANGELOG.md
@@ -7,8 +7,14 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Make the `ethProvider` constructor parameter optional, enabling access to read functions without requiring a wallet.
-- Changed `fetchMyContacts`, `fetchUserContacts`, and `sendEmail` to handle more precise errors
-- Removed `OriginalError` from `WorkflowError`
+- [BREAKING] Ship ES2023 JavaScript instead of es2015 (aka es6):
+  - Minimum browser versions: https://gist.github.com/Julien-Marcou/156b19aea4704e1d2f48adafc6e2acbf
+  - Minimum Node.js version: 18
+
+- Changed `fetchMyContacts`, `fetchUserContacts` and `sendEmail` error handling:
+  - Distinguish iExec protocol errors from other errors
+  - Store original error as the error cause
+- Removed `originalError` from `WorkflowError`
 - Upgrade typescript version
 - Move to target `ES2023` in order to support `cause` optional field in `Error`.